Change column types to bigint for owner and message in countrydex tables

+++ data/tables/spawn.sql
@@ -1,5 +1,5 @@
 create table _spawn (
-    message int primary key,
+    message bigint primary key,
     solved boolean,
     code varchar(6),
